fix(Emcee): fix error #9
Found error #9.
Error #9 is about printing double value with exponent.
Should print like 00.0% but prints 0.000E0.
Cause of error:
   - Double value are basically with exponent.
Fixed:
   - Use BigDecimal

Closed #9

+++ src/main/java/lotto/participant/Emcee.java
@@ -3,6 +3,7 @@ package lotto.participant;
 import lotto.domain.Lotto;
 import lotto.domain.Rank;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -64,6 +65,6 @@ public class Emcee {
     }
 
     public void showRateOfReturn(double rateOfReturn) {
-        System.out.println("총 수익률은 " + rateOfReturn + "%입니다.");
+        System.out.println("총 수익률은 " + BigDecimal.valueOf(rateOfReturn) + "%입니다.");
     }
 }
